+++ access_restricted/tests/__init__.py
@@ -1,2 +1,4 @@
 # -*- coding: utf-8 -*-
 from . import test_fields_view_get
+from . import test_fields_get
+from . import test_allow_implied